Move back integration tests into the example dir.
