--- examples/basic/README.md
@@ -1,28 +1,0 @@
-
-# Basic example
-
-## How to use
-
-Download the example (or clone the repo)[https://github.com/zeit/next.js.git]:
-
-```bash
-curl https://codeload.github.com/zeit/next.js/tar.gz/master | tar -xz --strip=2 next.js-master/examples/basic
-cd basic
-```
-
-Install it and run:
-
-```bash
-npm install
-npm run dev
-```
-
-Deploy it to the cloud with [now](https://zeit.co/now) ([download](https://zeit.co/download))
-
-```bash
-now
-```
-
-## The idea behind the example
-
-This example contains some of the basic features of Next.js. We use this example as a way to test basic features in Next.js
