--- examples/basic/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "basic",
-  "version": "1.0.0",
-  "scripts": {
-    "dev": "next",
-    "build": "next build",
-    "start": "next start"
-  },
-  "dependencies": {
-    "next": "*"
-  },
-  "author": "",
-  "license": "ISC"
-}
