+++ test/integration/basic/pages/async-props.js
