+++ test/integration/basic/pages/error.js
