+++ test/integration/basic/pages/finish-response.js
