+++ test/integration/basic/pages/head.js
