+++ test/integration/basic/pages/link.js
