+++ test/integration/basic/pages/stateful.js
