+++ test/integration/basic/pages/stateless.js
