+++ test/integration/basic/pages/styled-jsx.js
