+++ test/integration/basic/test/misc.js
