+++ test/integration/basic/test/rendering.js
